Refactor availableServices function and move it to index.ts
